Moved data from core to the root.
Refs: #109
