+++ firenado/conf/firenado.yml
@@ -14,9 +14,9 @@ components:
 data:
   connectors:
     - name: redis
-      class: firenado.core.data.connectors.RedisConnector
+      class: firenado.data.RedisConnector
     - name: sqlalchemy
-      class: firenado.core.data.connectors.SqlalchemyConnector
+      class: firenado.data.SqlalchemyConnector
 
 log:
   format: '%(levelname)s - %(name)s[%(module)s:%(lineno)d] - %(asctime)s - %(message)s'
